IPTE-149: Include a new Settings.class, code Refactoring
