+++ src/main/java/de/doubleslash/usb_led_matrix/usb_adapter/UsbAdapter.java
@@ -4,10 +4,10 @@ import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 
+import de.doubleslash.usb_led_matrix.CommandLineOptions;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import de.doubleslash.usb_led_matrix.Settings;
 import javafx.application.Platform;
 import javafx.beans.property.SimpleBooleanProperty;
 import javafx.beans.property.SimpleStringProperty;
@@ -20,7 +20,7 @@ import jssc.SerialPortList;
 
 public class UsbAdapter implements SerialPortEventListener {
 
-   byte numberLEDS = (byte) Settings.getNumberOfLeds();
+   byte numberLEDS = (byte) CommandLineOptions.getNumberOfLeds();
    private static final Logger LOG = LoggerFactory.getLogger(UsbAdapter.class);
 
    private final SimpleBooleanProperty connected = new SimpleBooleanProperty();
