+++ src/main/java/de/doubleslash/usb_led_matrix/view/ConfigurationView.java
@@ -1,6 +1,6 @@
 package de.doubleslash.usb_led_matrix.view;
 
-import de.doubleslash.usb_led_matrix.Settings;
+import de.doubleslash.usb_led_matrix.CommandLineOptions;
 import de.doubleslash.usb_led_matrix.graph.Graph;
 import de.doubleslash.usb_led_matrix.model.Model;
 import de.doubleslash.usb_led_matrix.resources.Resources;
@@ -238,7 +238,7 @@ public class ConfigurationView {
                      popUpStage.showAndWait();
                      startAfterAuthentification = true;
                   } catch (IOException e) {
-                     LOG.error("Could not load view '{}'. Exiting.", Resources.AUTHENTICATION_VIEW, e);
+                     LOG.error("Could not load view '{}'", Resources.AUTHENTICATION_VIEW, e);
                   }
                }
             }
@@ -292,7 +292,7 @@ public class ConfigurationView {
             model.setPortName(newValue);
          }
       });
-      selectPortComboBoxItem(Settings.getCom());
+      selectPortComboBoxItem(CommandLineOptions.getCom());
    }
 
    private void selectPortComboBoxItem(final String portName) {
@@ -331,16 +331,16 @@ public class ConfigurationView {
          timedOut = false;
       });
       this.usbAdapter = usbAdapter;
-      if (Settings.getColorMode().equals("light")) {
-         Settings.Light(scene);
-      } else if (Settings.getColorMode().equals("dark")) {
-         Settings.Dark(scene);
+      if (CommandLineOptions.getColorMode().equals("light")) {
+         CommandLineOptions.Light(scene);
+      } else if (CommandLineOptions.getColorMode().equals("dark")) {
+         CommandLineOptions.Dark(scene);
       }
       portChoiceBox.setItems(model.getSerialPorts());
       initializePortChoiceBox();
-      if (Settings.getMode().equals("manual")) {
+      if (CommandLineOptions.getMode().equals("manual")) {
          radioButtonGroup.selectToggle(manualRadioButton);
-      } else if (Settings.getMode().equals("teams")) {
+      } else if (CommandLineOptions.getMode().equals("teams")) {
          radioButtonGroup.selectToggle(teamsRadioButton);
       }
       if (model.getSelectedToggle() == null) {
@@ -349,8 +349,8 @@ public class ConfigurationView {
       } else {
          radioButtonGroup.selectToggle(model.getSelectedToggle());
       }
-      model.setBrightnessFromPercentage(Settings.getBrightness());
-      brightnessSlider.setValue(Settings.getBrightness());
+      model.setBrightnessFromPercentage(CommandLineOptions.getBrightness());
+      brightnessSlider.setValue(CommandLineOptions.getBrightness());
       registerConnectedListener();
       healthCheckThread = new Thread(healthCheckRunnable);
       healthCheckThread.setName("Health");
@@ -424,13 +424,13 @@ public class ConfigurationView {
       model.getSerialPorts().setAll(UsbAdapter.getSerialPortNames());
       if (model.getSerialPorts().contains(currentPort)) {
          selectPortComboBoxItem(currentPort);
-      } else if (model.getSerialPorts().contains(Settings.getCom())) {
-         selectPortComboBoxItem(Settings.getCom());
+      } else if (model.getSerialPorts().contains(CommandLineOptions.getCom())) {
+         selectPortComboBoxItem(CommandLineOptions.getCom());
       }
    }
 
    void turnOffAutomaticallyIfNeeded(final LocalTime currentTime) {
-      final LocalTime timeLightsTurnedOff = timeLightsTurnedOffNow.plusMinutes(Settings.getTimeout());
+      final LocalTime timeLightsTurnedOff = timeLightsTurnedOffNow.plusMinutes(CommandLineOptions.getTimeout());
       if (!currentTime.isBefore(timeLightsTurnedOff) && !timedOut) {
          LOG.info("Turn off automatically at '{}'", timeLightsTurnedOff);
          usbAdapter.updatePixel(Color.BLACK);
@@ -444,18 +444,20 @@ public class ConfigurationView {
 
    @FXML
    void toggleLightMode(final MouseEvent event) {
-      Settings.Light(scene);
+      CommandLineOptions.Light(scene);
    }
 
    @FXML
    void toggleDarkMode(final MouseEvent event) {
-      Settings.Dark(scene);
+      CommandLineOptions.Dark(scene);
    }
 
    @FXML
    void showInfoView() {
       final FXMLLoader fxmlLoader = new FXMLLoader(Resources.INFO_VIEW.getResource());
       try {
+         Stage stage = new Stage();
+
          usbAdapter.requestVersion();
          final Parent root = fxmlLoader.load();
          final VersionView versionView = fxmlLoader.getController();
@@ -463,14 +465,14 @@ public class ConfigurationView {
          versionView.setScene(scene);
          versionView.instantiate(usbAdapter.versionProperty());
 
-         popUpStage.setScene(scene);
-         popUpStage.getIcons().add(image);
-         popUpStage.setTitle("Info");
-         popUpStage.setResizable(false);
-         popUpStage.initModality(Modality.APPLICATION_MODAL);
-         popUpStage.show();
+         stage.setScene(scene);
+         stage.getIcons().add(image);
+         stage.setTitle("Info");
+         stage.setResizable(false);
+         stage.initModality(Modality.APPLICATION_MODAL);
+         stage.show();
       } catch (final IOException | SerialPortException e) {
-         LOG.error("Could not load view '{}'. Exiting.", Resources.INFO_VIEW, e);
+         LOG.error("Could not load view '{}'.", Resources.INFO_VIEW, e);
       }
    }
 
@@ -488,7 +490,7 @@ public class ConfigurationView {
          stage.initModality(Modality.APPLICATION_MODAL);
          stage.show();
       } catch (final IOException e) {
-         LOG.error("Could not load view '{}'. Exiting.", Resources.SETTINGS_VIEW, e);
+         LOG.error("Could not load view '{}'.", Resources.SETTINGS_VIEW, e);
       }
    }
 }
