+++ src/test/java/de/doubleslash/usb_led_matrix/CommandLineOptionsTest.java
@@ -11,109 +11,109 @@ import javafx.scene.paint.Color;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.EnumSource;
 
-class SettingsTest {
+class CommandLineOptionsTest {
    @BeforeEach
    void resetSettings() {
-      Settings.reset();
+      CommandLineOptions.reset();
    }
 
    @Test
    void shouldSetBrightnessValueWhenConfiguredWithLongCommandLineOption() {
       String[] args = { "--brightness", "60" };
-      Settings.handleParameters(args);
-      assertThat(Settings.getBrightness()).isEqualTo(60);
+      CommandLineOptions.handleParameters(args);
+      assertThat(CommandLineOptions.getBrightness()).isEqualTo(60);
    }
 
    @Test
    void shouldSetBrightnessValueWhenConfiguredWithShortCommandLineOption() {
       String[] args = { "-b", "60" };
-      Settings.handleParameters(args);
-      assertThat(Settings.getBrightness()).isEqualTo(60);
+      CommandLineOptions.handleParameters(args);
+      assertThat(CommandLineOptions.getBrightness()).isEqualTo(60);
    }
 
    @Test
    void shouldSetModeToManualWhenConfiguredWithLongCommandLineOption() {
       String[] args = { "--mode", "manual" };
-      Settings.handleParameters(args);
-      assertThat(Settings.getMode()).isEqualTo("manual");
+      CommandLineOptions.handleParameters(args);
+      assertThat(CommandLineOptions.getMode()).isEqualTo("manual");
    }
 
    @Test
    void shouldSetModeToManualWhenConfiguredWithShortCommandLineOption() {
       String[] args = { "-m", "manual" };
-      Settings.handleParameters(args);
-      assertThat(Settings.getMode()).isEqualTo("manual");
+      CommandLineOptions.handleParameters(args);
+      assertThat(CommandLineOptions.getMode()).isEqualTo("manual");
    }
 
    @Test
    void shouldSetColorModeToDarkWhenConfiguredWithLongCommandLineOption() {
       String[] args = { "--colorMode", "dark" };
-      Settings.handleParameters(args);
-      assertThat(Settings.getColorMode()).isEqualTo("dark");
+      CommandLineOptions.handleParameters(args);
+      assertThat(CommandLineOptions.getColorMode()).isEqualTo("dark");
    }
 
    @Test
    void shouldSetColorModeToDarkWhenConfiguredWithShortCommandLineOption() {
       String[] args = { "-c", "dark" };
-      Settings.handleParameters(args);
-      assertThat(Settings.getColorMode()).isEqualTo("dark");
+      CommandLineOptions.handleParameters(args);
+      assertThat(CommandLineOptions.getColorMode()).isEqualTo("dark");
    }
 
    @Test
    void shouldSetTimeoutWhenConfiguredWithLongCommandLineOption() {
       String[] args = { "--timeout", "30" };
-      Settings.handleParameters(args);
-      assertThat(Settings.getTimeout()).isEqualTo(30);
+      CommandLineOptions.handleParameters(args);
+      assertThat(CommandLineOptions.getTimeout()).isEqualTo(30);
    }
 
    @Test
    void shouldSetTimeoutWhenConfiguredWithShortCommandLineOption() {
       String[] args = { "-t", "30" };
-      Settings.handleParameters(args);
-      assertThat(Settings.getTimeout()).isEqualTo(30);
+      CommandLineOptions.handleParameters(args);
+      assertThat(CommandLineOptions.getTimeout()).isEqualTo(30);
    }
 
    @Test
    void shouldSetComPortWhenConfiguredWithLongCommandLineOption() {
       String[] args = { "--port", "COM7" };
-      Settings.handleParameters(args);
-      assertThat(Settings.getCom()).isEqualTo("COM7");
+      CommandLineOptions.handleParameters(args);
+      assertThat(CommandLineOptions.getCom()).isEqualTo("COM7");
    }
 
    @Test
    void shouldSetComPortWhenConfiguredWithShortCommandLineOption() {
       String[] args = { "-p", "COM7" };
-      Settings.handleParameters(args);
-      assertThat(Settings.getCom()).isEqualTo("COM7");
+      CommandLineOptions.handleParameters(args);
+      assertThat(CommandLineOptions.getCom()).isEqualTo("COM7");
    }
 
    @Test
    void shouldSetNumberOfLedsWhenConfiguredWithLongCommandLineOption() {
       String[] args = { "--numberOfLeds", "5" };
-      Settings.handleParameters(args);
-      assertThat(Settings.getNumberOfLeds()).isEqualTo(5);
+      CommandLineOptions.handleParameters(args);
+      assertThat(CommandLineOptions.getNumberOfLeds()).isEqualTo(5);
    }
 
    @Test
    void shouldSetNumberOfLedsWhenConfiguredWithShortCommandLineOption() {
       String[] args = { "-nr", "5" };
-      Settings.handleParameters(args);
-      assertThat(Settings.getNumberOfLeds()).isEqualTo(5);
+      CommandLineOptions.handleParameters(args);
+      assertThat(CommandLineOptions.getNumberOfLeds()).isEqualTo(5);
    }
 
    @Test
    void shouldNotSetColorModeToDarkWhenConfiguredWithInvalidValue() {
       String[] args = { "-c", "pink" };
-      Settings.handleParameters(args);
-      assertThat(Settings.getColorMode()).isNotEqualTo("pink");
+      CommandLineOptions.handleParameters(args);
+      assertThat(CommandLineOptions.getColorMode()).isNotEqualTo("pink");
    }
 
    @Test
    void shouldSetComPortAndNumberOfLedsWhenConfiguredWithShortCommandLineOption() {
       String[] args = { "-nr", "5", "-p", "COM7" };
-      Settings.handleParameters(args);
-      assertThat(Settings.getNumberOfLeds()).isEqualTo(5);
-      assertThat(Settings.getCom()).isEqualTo("COM7");
+      CommandLineOptions.handleParameters(args);
+      assertThat(CommandLineOptions.getNumberOfLeds()).isEqualTo(5);
+      assertThat(CommandLineOptions.getCom()).isEqualTo("COM7");
    }
 
    @ParameterizedTest
